refactor: update rsc, streaming ssr runtime code
