+++ packages/runtime/render/src/server/rsc/index.ts
@@ -1,1 +1,2 @@
 export * from './rsc';
+export * from './processRSCStream';
